Forecast temperatures and condition icons are correct from Google

+++ src/org/anddev/android/weatherforecast/weather/WeatherUtils.java
@@ -3,11 +3,11 @@ package org.anddev.android.weatherforecast.weather;
 /** Useful Utility in working with temperatures. (conversions). */
 public class WeatherUtils {
 
-	public static int fahrenheitToCelsius(int tFahrenheit) {
-		return (int) ((5.0f / 9.0f) * (tFahrenheit - 32));
+	public static Float fahrenheitToCelsius(float tFahrenheit) {
+		return ((5.0f / 9.0f) * (tFahrenheit - 32));
 	}
 
-	public static int celsiusToFahrenheit(int tCelsius) {
-		return (int) ((9.0f / 5.0f) * tCelsius + 32);
+	public static Float celsiusToFahrenheit(float tCelsius) {
+		return ((9.0f / 5.0f) * tCelsius + 32);
 	}
 }
